Use long for IDs, not int. Allow subclasses of objects in serialization... PreQuasiNodeTest: - Same as the datastore test, except: - If you ask it to fetch a key, it can ask its partner node to fetch it, if it's not in its own store. - If you ask it to put a key, it will still insert it only to its own store
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@7153 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/io/xfer/BlockReceiver.java
@@ -46,12 +46,12 @@ public class BlockReceiver {
 	public static final int MAX_SEND_INTERVAL = 500;
 	PartiallyReceivedBlock _prb;
 	Peer _sender;
-	int _uid;
+	long _uid;
 	UdpSocketManager _usm;
 	/** packet : Integer -> reportTime : Long * */
 	HashMap _recentlyReportedMissingPackets = new HashMap();
 
-	public BlockReceiver(UdpSocketManager usm, Peer sender, int uid, PartiallyReceivedBlock prb) {
+	public BlockReceiver(UdpSocketManager usm, Peer sender, long uid, PartiallyReceivedBlock prb) {
 		_sender = sender;
 		_prb = prb;
 		_uid = uid;
